refactor: extract the runner function from bin
