+++ bin/barnard59.js
@@ -4,15 +4,7 @@ const fs = require('fs')
 const p = require('..')
 const path = require('path')
 const program = require('commander')
-const cf = require('clownface')
-const rdf = require('rdf-ext')
-const namespace = require('@rdfjs/namespace')
-const TextLogStream = require('../lib/TextLogStream')
-
-const ns = {
-  p: namespace('https://pipeline.described.at/'),
-  rdf: namespace('http://www.w3.org/1999/02/22-rdf-syntax-ns#')
-}
+const runner = require('../lib/runner')
 
 function createOutputStream (output) {
   if (output === '-') {
@@ -32,32 +24,6 @@ function parseVariables (str, all) {
     }, all)
 }
 
-function guessPipeline (dataset) {
-  const graph = cf(dataset)
-
-  const pipelines = graph.has(ns.rdf('type'), [ ns.p('Pipeline'), ns.p('ObjectPipeline') ])
-
-  if (pipelines.values.length === 0) {
-    throw new Error('no pipeline found in the dataset')
-  }
-
-  const rootPipelines = pipelines.values.reduce((arr, id) => {
-    const node = dataset.match(null, null, rdf.namedNode(id), null)
-
-    if (node.length === 0) {
-      arr.push(id)
-    }
-
-    return arr
-  }, [])
-
-  if (rootPipelines.length > 1) {
-    throw new Error('multiple root pipeline found. please specify the one to run using --pipeline option')
-  }
-
-  return rootPipelines[0]
-}
-
 program
   .command('run <filename>')
   .option('--format <mediaType>', 'media type of the pipeline description', 'application/ld+json')
@@ -65,47 +31,16 @@ program
   .option('--pipeline [iri]', 'IRI of the pipeline description')
   .option('--variable <name=value>', 'variable key value pairs separated by comma', parseVariables, new Map())
   .option('-v, --verbose', 'enable diagnostic console output')
-  .action((filename, { format, output, pipeline, variable, verbose } = {}) => {
+  .action((filename, options = {}) => {
+    const { format, output } = options
     p.fileToDataset(format, filename)
-      .then(dataset => {
-        if (!pipeline) {
-          pipeline = guessPipeline(dataset)
-        }
-
-        if (verbose) {
-          process.stdout.write('variables via CLI:\n')
-
-          for (const [key, value] of variable) {
-            process.stdout.write(`  ${key}: ${value}\n`)
-          }
-        }
-
-        const stream = p.pipeline(dataset, pipeline, {
-          basePath: path.resolve(path.dirname(filename)),
-          variables: variable
-        })
-
-        if (verbose) {
-          process.stdout.write('variables in pipeline instance:\n')
-
-          for (const [key, value] of stream.variables) {
-            process.stdout.write(`  ${key}: ${value}\n`)
-          }
-        }
-
-        stream.on('error', err => {
-          console.error(err)
-          process.exit(1)
-        })
-
-        stream.pipe(createOutputStream(output))
-
-        if (verbose) {
-          stream.context.log.pipe(new TextLogStream()).pipe(process.stdout)
-        }
-
-        return p.run(stream)
-      }).catch(err => {
+      .then(runner.create({
+        ...options,
+        outputStream: createOutputStream(output),
+        log: process.stdout,
+        basePath: path.resolve(path.dirname(filename))
+      }))
+      .catch(err => {
         console.error(err)
         process.exit(1)
       })
